requirements-dev.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.5
boto3==1.42.88
awscrt==0.28.1
awsiotsdk==1.26.0
flask==3.1.2
